requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2026.1.4
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
dash==3.3.0
dash-bootstrap-components==2.0.4
dash-iconify==0.1.2
dash_mantine_components==2.4.1
Django==6.0.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mptt==0.18.0
django-mptt-admin==2.9.0
django-plotly-dash==2.5.0
dotenv==0.9.9
dpd_components==0.2.0
et_xmlfile==2.0.0
Flask==3.1.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
plotly==6.5.2
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
retrying==1.4.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
whitenoise==6.11.0
zipp==3.23.0
